Move everything out of the old core libs, and fix up
