+++ packages/app/src/App.tsx
@@ -20,7 +20,6 @@ import {
   FlatRoutes,
   OAuthRequestDialog,
   SignInPage,
-  ErrorBoundaryFallback,
 } from '@backstage/core';
 import { apiDocsPlugin, ApiExplorerPage } from '@backstage/plugin-api-docs';
 import {
@@ -67,9 +66,6 @@ const app = createApp({
     alert: AlarmIcon,
   },
   components: {
-    ErrorBoundaryFallback: props => {
-      return <ErrorBoundaryFallback {...props} />;
-    },
     SignInPage: props => {
       return (
         <SignInPage
