+++ packages/core-api/src/extensions/extensions.test.tsx
@@ -14,13 +14,7 @@
  * limitations under the License.
  */
 
-import { errorApiRef } from '../apis/definitions/ErrorApi';
-import { ApiRegistry } from '../apis/system/ApiRegistry';
-import { ApiProvider } from '../apis/system/ApiProvider';
-import { withLogCollector } from '@backstage/test-utils-core';
-import { render, screen } from '@testing-library/react';
 import React from 'react';
-import { AppContextProvider } from '../app/AppContext';
 import { createPlugin } from '../plugin';
 import { createRouteRef } from '../routing';
 import { getComponentData } from './componentData';
@@ -29,9 +23,6 @@ import {
   createReactExtension,
   createRoutableExtension,
 } from './extensions';
-import { AppComponents } from '../app/types';
-import { AppContext } from '../app';
-import { MockErrorApi } from '@backstage/test-utils';
 
 const plugin = createPlugin({
   id: 'my-plugin',
@@ -81,45 +72,5 @@ describe('extensions', () => {
     expect(getComponentData(element1, 'core.plugin')).toBe(plugin);
     expect(getComponentData(element2, 'core.plugin')).toBe(plugin);
     expect(getComponentData(element2, 'core.mountPoint')).toBe(routeRef);
-  });
-
-  it('should wrap extended component with error boundary', async () => {
-    const extension = createComponentExtension({
-      component: {
-        sync: () => {
-          throw new Error('Test error');
-        },
-      },
-    });
-
-    const BrokenComponent = plugin.provide(extension);
-
-    const errorApi = new MockErrorApi();
-
-    const apis = ApiRegistry.from([[errorApiRef, errorApi]]);
-
-    const MockFallback: AppComponents['ErrorBoundaryFallback'] = props => (
-      <>Error in {props.plugin?.getId()}</>
-    );
-
-    const { error: errors } = await withLogCollector(['error'], async () => {
-      render(
-        <ApiProvider apis={apis}>
-          <AppContextProvider
-            appContext={
-              ({
-                getComponents: () => ({
-                  ErrorBoundaryFallback: MockFallback,
-                }),
-              } as unknown) as AppContext
-            }
-          >
-            <BrokenComponent />
-          </AppContextProvider>
-        </ApiProvider>,
-      );
-    });
-    screen.getByText('Error in my-plugin');
-    expect(errors[0]).toMatch('Test error');
   });
 });
