+++ packages/core-app-api/src/app/App.test.tsx
@@ -159,6 +159,7 @@ describe('Integration Test', () => {
     BootErrorPage: () => null,
     Progress: () => null,
     Router: BrowserRouter,
+    ErrorBoundaryFallback: () => null,
   };
 
   it('runs happy paths', async () => {
