+++ packages/core-components/src/components/ErrorPanel/index.ts
