+++ packages/core-components/src/components/ResponseErrorPanel/index.ts
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-export { ResponseErrorDetails, ResponseErrorPanel } from './ResponseErrorPanel';
+export { ResponseErrorPanel } from './ResponseErrorPanel';
