+++ packages/core-components/src/components/index.ts
@@ -22,6 +22,7 @@ export * from './CopyTextButton';
 export * from './DependencyGraph';
 export * from './DismissableBanner';
 export * from './EmptyState';
+export * from './ErrorPanel';
 export * from './ResponseErrorPanel';
 export * from './FeatureDiscovery';
 export * from './HeaderIconLinkRow';
