+++ packages/core/src/api-wrappers/createApp.tsx
@@ -30,7 +30,6 @@ import LightIcon from '@material-ui/icons/WbSunny';
 import DarkIcon from '@material-ui/icons/Brightness2';
 import { ErrorPage } from '../layout/ErrorPage';
 import { Progress } from '../components/Progress';
-import { ErrorBoundaryFallback } from '../components/ErrorBoundaryFallback';
 import { defaultApis } from './defaultApis';
 import { lightTheme, darkTheme } from '@backstage/theme';
 import { AppConfig, JsonObject } from '@backstage/config';
@@ -131,7 +130,6 @@ export function createApp(options?: AppOptions) {
     BootErrorPage: DefaultBootErrorPage,
     Progress: Progress,
     Router: BrowserRouter,
-    ErrorBoundaryFallback: ErrorBoundaryFallback,
     ...options?.components,
   };
   const themes = options?.themes ?? [
