+++ packages/test-utils/src/testUtils/appWrappers.tsx
@@ -96,7 +96,6 @@ export function wrapInTestApp(
       Router: ({ children }) => (
         <MemoryRouter initialEntries={routeEntries} children={children} />
       ),
-      ErrorBoundaryFallback: () => null,
     },
     icons: defaultSystemIcons,
     plugins: [],
